requirements.txt
Django
djangorestframework
psycopg2-binary 
gunicorn